requirements.txt
pytest
pydantic
requests
allure-pytest
fastapi
